+++ src/Cleff/Internal/Stack.hs
@@ -8,8 +8,8 @@
 -- Stability: unstable
 -- Portability: non-portable (GHC only)
 --
--- This module defines an immutable extensible record type, similar to @vinyl@ and @data-diverse@. However this
--- implementation focuses on fast reads, hence has very different performance characteristics from other libraries:
+-- This module defines the effect stack as an immutable extensible stack type, and provides functions for manipulating
+-- it. The effect stack type has the following time complexities:
 --
 -- * Lookup: Amortized \( O(1) \).
 -- * Update: \( O(n) \).
@@ -18,9 +18,9 @@
 --
 -- __This is an /internal/ module and its API may change even between minor versions.__ Therefore you should be
 -- extra careful if you're to depend on this module.
-module Cleff.Internal.Rec
+module Cleff.Internal.Stack
   ( Effect
-  , Rec
+  , Stack
   , HandlerPtr (HandlerPtr, unHandlerPtr)
     -- * Construction
   , type (++)
@@ -53,10 +53,10 @@ import           Prelude      hiding (concat, drop, head, tail, take)
 -- @a :: 'Type'@.
 type Effect = (Type -> Type) -> Type -> Type
 
--- | Extensible record type supporting efficient \( O(1) \) reads. The underlying implementation is 'PrimArray'
--- slices.
-type role Rec nominal
-newtype Rec (es :: [Effect]) = Rec (PrimVec Int)
+-- | The effect stack, storing pointers to handlers. It is essentially an extensible stack type supporting
+-- efficient \( O(1) \) reads.
+type role Stack nominal
+newtype Stack (es :: [Effect]) = Stack (PrimVec Int)
 
 -- | A pointer to an effect handler.
 type role HandlerPtr nominal
@@ -68,12 +68,12 @@ unreifiable clsName funName comp = error $
   \to define an instance for the '" <> clsName <> "' typeclass, which you should not be doing whatsoever. If that or \
   \other shenanigans seem unlikely, please report this as a bug."
 
--- | Create an empty record. \( O(1) \).
-empty :: Rec '[]
+-- | Create an empty stack. \( O(1) \).
+empty :: Stack '[]
 empty = coerce (Vec.empty @Int)
 
--- | Prepend one entry to the record. \( O(n) \).
-cons :: HandlerPtr e -> Rec es -> Rec (e : es)
+-- | Prepend one entry to the stack. \( O(n) \).
+cons :: HandlerPtr e -> Stack es -> Stack (e : es)
 cons = coerce (Vec.cons @Int)
 
 -- | Type level list concatenation.
@@ -82,12 +82,12 @@ type family xs ++ ys where
   (x : xs) ++ ys = x : (xs ++ ys)
 infixr 5 ++
 
--- | Concatenate two records. \( O(m+n) \).
-concat :: Rec es -> Rec es' -> Rec (es ++ es')
+-- | Concatenate two stacks. \( O(m+n) \).
+concat :: Stack es -> Stack es' -> Stack (es ++ es')
 concat = coerce (Vec.concat @Int)
 
--- | Slice off one entry from the top of the record. \( O(1) \).
-tail :: Rec (e : es) -> Rec es
+-- | Slice off one entry from the top of the stack. \( O(1) \).
+tail :: Stack (e : es) -> Stack es
 tail = coerce (Vec.tail @Int)
 
 -- | @'KnownList' es@ means the list @es@ is concrete, /i.e./ is of the form @'[a1, a2, ..., an]@ instead of a type
@@ -95,7 +95,7 @@ tail = coerce (Vec.tail @Int)
 class KnownList (es :: [Effect]) where
   -- | Get the length of the list.
   reifyLen :: Int
-  reifyLen = unreifiable "KnownList" "Cleff.Internal.Rec.reifyLen" "the length of a type-level list"
+  reifyLen = unreifiable "KnownList" "Cleff.Internal.Stack.reifyLen" "the length of a type-level list"
 
 instance KnownList '[] where
   reifyLen = 0
@@ -103,16 +103,16 @@ instance KnownList '[] where
 instance KnownList es => KnownList (e : es) where
   reifyLen = 1 + reifyLen @es
 
--- | Slice off several entries from the top of the record. \( O(1) \).
-drop :: ∀ es es'. KnownList es => Rec (es ++ es') -> Rec es'
+-- | Slice off several entries from the top of the stack. \( O(1) \).
+drop :: ∀ es es'. KnownList es => Stack (es ++ es') -> Stack es'
 drop = coerce (Vec.drop @Int) (reifyLen @es)
 
--- | Get the head of the record. \( O(1) \).
-head :: Rec (e : es) -> HandlerPtr e
+-- | Get the head of the stack. \( O(1) \).
+head :: Stack (e : es) -> HandlerPtr e
 head = coerce (Vec.head @Int)
 
--- | Take elements from the top of the record. \( O(m) \).
-take :: ∀ es es'. KnownList es => Rec (es ++ es') -> Rec es
+-- | Take elements from the top of the stack. \( O(m) \).
+take :: ∀ es es'. KnownList es => Stack (es ++ es') -> Stack es
 take = coerce (Vec.take @Int) (reifyLen @es)
 
 -- | @e ':>' es@ means the effect @e@ is present in the effect stack @es@, and therefore can be 'Cleff.send'ed in an
@@ -120,7 +120,7 @@ take = coerce (Vec.take @Int) (reifyLen @es)
 class (e :: Effect) :> (es :: [Effect]) where
   -- | Get the index of the element.
   reifyIndex :: Int
-  reifyIndex = unreifiable "Elem" "Cleff.Internal.Rec.reifyIndex" "the index of an effect in the effect stack"
+  reifyIndex = unreifiable "Elem" "Cleff.Internal.Stack.reifyIndex" "the index of an effect in the effect stack"
 infix 0 :>
 
 -- | The element closer to the head takes priority.
@@ -134,7 +134,7 @@ type ElemNotFound e = Text "The element '" :<>: ShowType e :<>: Text "' is not p
 
 instance TypeError (ElemNotFound e) => e :> '[] where
   reifyIndex = error
-    "Cleff.Internal.reifyIndex: Attempting to refer to a nonexistent member. Please report this as a bug."
+    "Cleff.Internal.Stack.reifyIndex: Attempting to refer to a nonexistent member. Please report this as a bug."
 
 -- | @xs ':>>' es@ means the list of effects @xs@ are all present in the effect stack @es@. This is a convenient type
 -- alias for @(e1 ':>' es, ..., en ':>' es)@.
@@ -143,8 +143,8 @@ type family xs :>> es :: Constraint where
   (x : xs) :>> es = (x :> es, xs :>> es)
 infix 0 :>>
 
--- | Get an element in the record. Amortized \( O(1) \).
-index :: ∀ e es. e :> es => Rec es -> HandlerPtr e
+-- | Get an element in the stack. Amortized \( O(1) \).
+index :: ∀ e es. e :> es => Stack es -> HandlerPtr e
 index = coerce (Vec.index @Int) (reifyIndex @e @es)
 
 -- | @es@ is a subset of @es'@, /i.e./ all elements of @es@ are in @es'@.
@@ -152,7 +152,7 @@ class KnownList es => Subset (es :: [Effect]) (es' :: [Effect]) where
   -- | Get a list of indices of the elements.
   reifyIndices :: [Int]
   reifyIndices = unreifiable
-    "Subset" "Cleff.Internal.Rec.reifyIndices" "the indices of a subset of the effect stack"
+    "Subset" "Cleff.Internal.Stack.reifyIndices" "the indices of a subset of the effect stack"
 
 instance Subset '[] es where
   reifyIndices = []
@@ -160,10 +160,10 @@ instance Subset '[] es where
 instance (Subset es es', e :> es') => Subset (e : es) es' where
   reifyIndices = reifyIndex @e @es' : reifyIndices @es @es'
 
--- | Get a subset of the record. Amortized \( O(m) \).
-pick :: ∀ es es'. Subset es es' => Rec es' -> Rec es
+-- | Get a subset of the stack. Amortized \( O(m) \).
+pick :: ∀ es es'. Subset es es' => Stack es' -> Stack es
 pick = coerce (Vec.pick @Int) (reifyLen @es) (reifyIndices @es @es')
 
--- | Update an entry in the record. \( O(n) \).
-update :: ∀ e es. e :> es => HandlerPtr e -> Rec es -> Rec es
+-- | Update an entry in the stack. \( O(n) \).
+update :: ∀ e es. e :> es => HandlerPtr e -> Stack es -> Stack es
 update = coerce (Vec.update @Int) (reifyIndex @e @es)